BUG: Modify the string builder to handle 64 bit integers when adding to the string and not int32. The public comment makes no reference to the integer size. However the internal comment shows it was intended to be int 64 type. This change should not impact existing calls that pass int 32 as they should simply case to a int 64. As the method adds a base 10 string for the integer to the builder there should be no difference to the output.

+++ string.h
@@ -265,7 +265,7 @@ EXTERNAL fiftyoneDegreesStringBuilder* fiftyoneDegreesStringBuilderAddChar(
  */
 EXTERNAL fiftyoneDegreesStringBuilder* fiftyoneDegreesStringBuilderAddInteger(
 	fiftyoneDegreesStringBuilder* builder,
-	int const value);
+	int64_t const value);
 
 /**
  * Adds the string to the buffer.
